+++ extensions/wave/CHANGELOG.md
@@ -1,3 +1,8 @@
 # Wave Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- View Wave Businesses
+    - View Invoices
+    - View Customers
+    - View (some) Customer Statements